Planet_Procedural_Generation Demo AddOn
This is "Planet_Procedural_Generation 0.8.3" demo file created to work as addon for your blender so you wouldn't need to go through scripting menu. All you need to do is download it, zip it and then in Blender go to -> Edit -> Preferences -> Add-ons -> Install from Disk... -> and select this file in .zip.